Fixes with JAX update

+++ jax_elo/utils/linalg.py
@@ -1,5 +1,8 @@
 import jax.numpy as jnp
 from jax.ops import index_update
+from jax import jit
+from functools import partial
+import numpy as np
 
 
 def pos_def_mat_from_tri_elts(elts, mat_size, jitter=1e-6):
@@ -14,9 +17,9 @@ def pos_def_mat_from_tri_elts(elts, mat_size, jitter=1e-6):
 
 def num_mat_elts(num_triangular_elts):
 
-    sqrt_term = jnp.sqrt(8 * num_triangular_elts + 1)
+    sqrt_term = np.sqrt(8 * num_triangular_elts + 1)
 
-    return int((sqrt_term - 1) / 2)
+    return ((sqrt_term - 1) / 2).astype(int)
 
 
 def num_triangular_elts(mat_size, include_diagonal=True):
